+++ src/main/java/hellofx/fheroes2/objects/ObjTree.java
@@ -1,7 +1,7 @@
 package hellofx.fheroes2.objects;
 
 public class ObjTree {
-    public static int GetPassable(byte index) {
+    public static int GetPassable(int index) {
         //TODO
         return 0;
     }
